Refactor FHIR complex type compaction
