+++ src/fhir_mcp_server/compactor/compactors/money.py
@@ -1,0 +1,32 @@
+# Copyright (c) 2026, WSO2 LLC. (https://www.wso2.com/) All Rights Reserved.
+
+# WSO2 LLC. licenses this file to you under the Apache License,
+# Version 2.0 (the "License"); you may not use this file except
+# in compliance with the License.
+# You may obtain a copy of the License at
+
+# http://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied. See the License for the
+# specific language governing permissions and limitations
+# under the License.
+
+import logging
+
+from ..types.complex_types import Money
+
+logger = logging.getLogger(__name__)
+
+
+def compact(money: Money) -> str:
+    """Compact a Money value to "amount currency"."""
+    # {"value": 49.99, "currency": "USD"} -> "49.99 USD"
+    logger.debug(f"Compacting Money: {money}")
+    value = f"{float(money.value):g}" if money.value is not None else ""
+    currency = money.currency or ""
+    compacted = f"{value} {currency}".strip()
+    logger.debug(f"Compacted Money: '{compacted}'")
+    return compacted
