+++ src/fhir_mcp_server/compactor/registry.py
@@ -1,0 +1,56 @@
+# Copyright (c) 2026, WSO2 LLC. (https://www.wso2.com/) All Rights Reserved.
+
+# WSO2 LLC. licenses this file to you under the Apache License,
+# Version 2.0 (the "License"); you may not use this file except
+# in compliance with the License.
+# You may obtain a copy of the License at
+
+# http://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied. See the License for the
+# specific language governing permissions and limitations
+# under the License.
+
+from .compactors import (
+    address,
+    annotation,
+    attachment,
+    codeable_concept,
+    coding,
+    contact_point,
+    extension,
+    human_name,
+    identifier,
+    money,
+    period,
+    quantity,
+    range as range_,
+    ratio,
+    timing,
+)
+from .types.complex_types import (
+    Address, Annotation, Attachment, CodeableConcept, Coding,
+    ContactPoint, Extension, HumanName, Identifier, Money,
+    Period, Quantity, Range, Ratio, Timing,
+)
+
+COMPACTORS = {
+    Ratio: ratio.compact,
+    Range: range_.compact,
+    Money: money.compact,
+    Annotation: annotation.compact,
+    Period: period.compact,
+    HumanName: human_name.compact,
+    Address: address.compact,
+    Coding: coding.compact,
+    CodeableConcept: codeable_concept.compact,
+    Quantity: quantity.compact,
+    Extension: extension.compact,
+    Attachment: attachment.compact,
+    ContactPoint: contact_point.compact,
+    Identifier: identifier.compact,
+    Timing: timing.compact,
+}
