+++ src/fhir_mcp_server/compactor/types/complex_types/quantity.py
@@ -1,0 +1,28 @@
+# Copyright (c) 2026, WSO2 LLC. (https://www.wso2.com/) All Rights Reserved.
+
+# WSO2 LLC. licenses this file to you under the Apache License,
+# Version 2.0 (the "License"); you may not use this file except
+# in compliance with the License.
+# You may obtain a copy of the License at
+
+# http://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied. See the License for the
+# specific language governing permissions and limitations
+# under the License.
+
+
+from typing import Optional, Union
+
+from .base import FhirBaseModel
+
+
+class Quantity(FhirBaseModel):
+    value: Optional[Union[int, float]] = None
+    comparator: Optional[str] = None
+    unit: Optional[str] = None
+    system: Optional[str] = None
+    code: Optional[str] = None
